+++ src/components/MainListInfoForm/VisibilityTooltip/index.ts
@@ -1,0 +1,1 @@
+export { VisibilityTooltip } from './VIsibilityTooltip';